Moved all OStatus code in a single file.
